Refactored files and fs
