+++ lib/virtual-fs/src/files/_static.rs
