+++ lib/virtual-fs/src/files/arc.rs
