+++ lib/virtual-fs/src/files/arc_box.rs
