+++ lib/virtual-fs/src/files/combine.rs
@@ -1,6 +1,6 @@
 use derivative::Derivative;
 
-use super::*;
+use crate::*;
 
 use crate::VirtualFile;
