+++ lib/virtual-fs/src/files/dual_write.rs
@@ -1,6 +1,6 @@
 use derivative::Derivative;
 
-use super::*;
+use crate::*;
 
 use crate::VirtualFile;
