+++ lib/virtual-fs/src/files/mod.rs
@@ -1,0 +1,11 @@
+pub mod _static;
+pub mod arc;
+pub mod arc_box;
+pub mod buffer;
+pub mod combine;
+pub mod cow;
+pub mod dual_write;
+pub mod null;
+pub mod random;
+pub mod special;
+pub mod zero;
