+++ lib/virtual-fs/src/files/special.rs
