+++ lib/virtual-fs/src/fs/_static.rs
