+++ lib/virtual-fs/src/fs/empty.rs
