+++ lib/virtual-fs/src/fs/overlay.rs
@@ -1059,7 +1059,7 @@ mod tests {
     use super::*;
     use crate::{mem_fs::FileSystem as MemFS, webc_fs::WebcFileSystem, RootFileSystemBuilder};
 
-    const PYTHON: &[u8] = include_bytes!("../../c-api/examples/assets/python-0.1.0.wasmer");
+    const PYTHON: &[u8] = include_bytes!("../../../c-api/examples/assets/python-0.1.0.wasmer");
 
     #[tokio::test]
     async fn remove_directory() {
